+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `environmentId` to point address element
+- `ADDRESS_PROPERTY` to conditional predicates schema
 
 ## [6.6.0] - 2024-06-13